+++ packages/design-system/.storybook/main.js
@@ -4,7 +4,18 @@ module.exports = {
     "../src/stories/Pages/*.stories.@(js|jsx|ts|tsx|mdx)",
     "../src/**/*.stories.@(js|jsx|ts|tsx|mdx)",
   ],
-  addons: ["@storybook/addon-links", "@storybook/addon-essentials"],
+  addons: [
+    "@storybook/addon-links",
+    "@storybook/addon-essentials",
+    {
+      name: "@storybook/preset-scss",
+      options: {
+        cssLoaderOptions: {
+          modules: true,
+        },
+      },
+    },
+  ],
   typescript: {
     reactDocgen: "react-docgen-typescript",
     reactDocgenTypescriptOptions: {