+++ packages/design-system/src/components/GlobalStyle/GlobalStyle.ts
@@ -16,11 +16,11 @@
 // =============================================================================
 import { createGlobalStyle } from "styled-components";
 import { rem } from "polished";
-import { palette, spacing } from "../../styles";
-import "@fontsource/public-sans";
+import { palette, spacing, typography } from "../../styles";
+import "@fontsource/public-sans/400.css";
 import "@fontsource/public-sans/500.css";
-import "@fontsource/libre-franklin";
-import "@fontsource/libre-franklin/500.css";
+import "@fontsource/public-sans/600.css";
+import "@fontsource/public-sans/700.css";
 import "@fontsource/libre-baskerville";
 
 export const GlobalStyle = createGlobalStyle`
@@ -45,20 +45,17 @@ export const GlobalStyle = createGlobalStyle`
   }
    
   body {
+    ${typography.Sans16}
     height: 100%;
     margin: 0;
     padding: 0;
     background: ${palette.marble3};
     color: ${palette.pine3};
-    font-family: ${(props) => props.theme.fonts.body};
-    font-size: ${rem("16px")};
-    letter-spacing: -0.01em;
-    line-height: 1.5;
     -webkit-font-smoothing: antialised;
     text-rendering: optimizeLegibility;
   }
   
   strong {
-    font-weight: 500;
+    font-weight: 700;
   }
 `;