+++ packages/design-system/src/components/Typography/index.ts
@@ -14,5 +14,6 @@
 // You should have received a copy of the GNU General Public License
 // along with this program.  If not, see <https://www.gnu.org/licenses/>.
 // =============================================================================
-export * from "./Headings";
 export * from "./Link";
+export * from "./UI";
+export * from "./Article";