+++ packages/design-system/src/styles/index.ts
@@ -15,7 +15,7 @@
 // along with this program.  If not, see <https://www.gnu.org/licenses/>.
 // =============================================================================
 export * from "./animation";
-export * from "./fonts";
 export * from "./palette";
 export * from "./spacing";
 export * from "./zindex";
+export * from "./typography";